1.0 dev (#23)
* Added a '#include <vector>' to repeat.hpp - without that there is a build error using Window's linux environment. Also starting overlap correction groundwork.

* Reformatted repeats_overlap(...)

* Change the implementation location of repeats_overlap

* Put in ground work for join_repeat_region(...)

* joint_repeat_region update: added sequence string joining

* joint_repeat_region update: added score joining

* Joining repeat splits is partially done

* joint_repeat_region(...) is done (for now).
In the future the function should just join the sequence + scores + logoNumbers and then rerun repeat splitting. We don't do that for runtime reasons (you would have to rerun repeat splitting in repeat output, or perhaps just do it only there). Reconsider in the future.

* Improved the bound on maxReadID in ultra.cpp

* Maybe repeat joining is actually working? This is the one to test

* Now with consensus name ordering.

* Debugged consensus ordering

* Debugged some repeat consensus stuff

* Debugged some repeat merging stuff. Seems to be working.

* Completed a TODO

* Splitting/merging complete...?

* Adding code for masking

* Adding code for masking

* We are now not allowing sequence starts to be ';'

* Mor masking code

* Changing masking defaults

* ran format

* Some cleanup + I think we got masking working... checking...

* Masking is mostly working. Need to make some adjustments to how we index masks

* Masking is fully functional

* Adding in new settings/CLI

* Added CLI11 code

* Adding in the memory calculation function (and also automatic memory assignment)

* Changed doubles to floats.

* New settings file in place and release default added to cmake file

* Everything is in place - except now there is a consensus bug. Some debugging printfs in this version

* Debugged the consensus bug

* We now have the json string out

* Update README.md

* We now have the json string out

* Added a '#include <vector>' to repeat.hpp - without that there is a build error using Window's linux environment. Also starting overlap correction groundwork.

* Reformatted repeats_overlap(...)

* Change the implementation location of repeats_overlap

* Put in ground work for join_repeat_region(...)

* joint_repeat_region update: added sequence string joining

* joint_repeat_region update: added score joining

* Joining repeat splits is partially done

* joint_repeat_region(...) is done (for now).
In the future the function should just join the sequence + scores + logoNumbers and then rerun repeat splitting. We don't do that for runtime reasons (you would have to rerun repeat splitting in repeat output, or perhaps just do it only there). Reconsider in the future.

* Improved the bound on maxReadID in ultra.cpp

* Maybe repeat joining is actually working? This is the one to test

* Now with consensus name ordering.

* Debugged consensus ordering

* Debugged some repeat consensus stuff

* Debugged some repeat merging stuff. Seems to be working.

* Completed a TODO

* Splitting/merging complete...?

* Adding code for masking

* Adding code for masking

* We are now not allowing sequence starts to be ';'

* Mor masking code

* Changing masking defaults

* ran format

* Some cleanup + I think we got masking working... checking...

* Masking is mostly working. Need to make some adjustments to how we index masks

* Masking is fully functional

* Adding in new settings/CLI

* Added CLI11 code

* Adding in the memory calculation function (and also automatic memory assignment)

* Changed doubles to floats.

* New settings file in place and release default added to cmake file

* Everything is in place - except now there is a consensus bug. Some debugging printfs in this version

* Debugged the consensus bug

* We now have the json string out

* Update README.md

* We now have the json string out

* Including <algorithm. in mask.h for GCC

* Removing check-format

Removing check-format

* Update README.md

* updating build yaml

* Update build-and-test.yml

* Finally got check/run-format to work. –.–

* Update build-and-test.yml

---------

+++ src/umatrix.cpp
@@ -78,7 +78,7 @@ void UMatrix::AdjustScoreToZero(int startingAt, int numberOfColumns) {
 }
 
 // Adjust every row in <column> to be [row] = [row] + adjustment
-void UMatrix::AdjustColumn(int column, double adjustment) {
+void UMatrix::AdjustColumn(int column, float adjustment) {
   for (int i = 0; i < cellsPerColumn; ++i) {
     scoreColumns[column][i] += adjustment;
   }
@@ -86,8 +86,8 @@ void UMatrix::AdjustColumn(int column, double adjustment) {
 
 // Return the score in scoreColumns[currentColumnIndex - d][row]
 // Account for warpping if currentColumnIndex - d < 0
-double UMatrix::PreviousScore(int row, int d) {
-  double s = NEG_INF;
+float UMatrix::PreviousScore(int row, int d) {
+  float s = NEG_INF;
 
   if (d < length) {
     int c = PreviousColumn(d);
@@ -411,10 +411,9 @@ void UMatrix::CreateMatrix() {
    }*/
 
   // Allocate matrices
-  scoreMatrix =
-      (double *)malloc(sizeof(double) * cellsPerColumn * (length + 1));
+  scoreMatrix = (float *)malloc(sizeof(float) * cellsPerColumn * (length + 1));
   tracebackMatrix = (int *)malloc(sizeof(int) * (maxPeriod + 1) * (length + 1));
-  scoreColumns = (double **)malloc(sizeof(double *) * (length + 1));
+  scoreColumns = (float **)malloc(sizeof(float *) * (length + 1));
   tracebackColumns = (int **)malloc(sizeof(int *) * (length + 1));
 
   // Assign columns of matrices
